Weekly Edit
I finished Lesson 2 and 3. I also used Nida's login code and merged it into the main file.

Most important thing now is to get the AI setup.

Next TODO is to get the login information able to be sent to the data base. We also still need to setup a function where it will jump to the users last module.